requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
groq
pillow
google-generativeai
requests
python-multipart
typing
azure-cognitiveservices-speech